Keep each test database in a different path; clean up afterwards
SQLite's DB.CLose() appears to be racy (! - I have no other plausible
explanation for this). Reusing the same database directory was leading
to occasional test failures for no good reason (DB creation failed
with a read-only DB).

This is a small refactor which should make our tests robust against this.