Handle empty InvokeStageResponse.
This stops functions configured with type "async" crashing the Flow
service.

The unmarshaller is permissive enough that an empty response from the
invoke stage call will create an empty `InvokeStageResponse`. The
invoke stage call will create an empty `InvokeStageResponse`
struct. This has a null result field, causing the Flow service to crash.

This change adds a check for a null result and causes the stage to
fail in the same way as when the response is non-empty but invalid.